User and project controllers finished (#42)
* Separated admin and grapesStorage routes and controller

* utilsController created

* Set up MongoDB

* Urls resolved

* Created userController.js file

* Updated userRouter.js

* Project router and controller created
